加载点云后就进行放大变换，将阈值改为double类型，放大点云每次依据原始点云进行放大

+++ common/Transfer_Function.cpp
@@ -18,7 +18,7 @@ void Transfer_Function::Cloud2cvMat(int width,int height,float originX,float ori
     }
 }
 
-void Transfer_Function::cvMat2Cloud(int minHeight, int maxHeight, cv::Mat& imageIn, pcl::PointCloud<pcl::PointXYZ>::Ptr cloudOut)
+void Transfer_Function::cvMat2Cloud(double minHeight, double maxHeight, cv::Mat& imageIn, pcl::PointCloud<pcl::PointXYZ>::Ptr cloudOut)
 {
     cloudOut->points.clear();
