+++ common/Transfer_Function.h
@@ -37,7 +37,7 @@ public:
 	/// <param name="maxHeight"></param>
 	/// <param name="imageIn">待转换的cv::Mat</param>
 	/// <param name="cloudOut">转换后的点云</param>
-	static void cvMat2Cloud(int minHeight,int maxHeight,cv::Mat& imageIn, pcl::PointCloud<pcl::PointXYZ>::Ptr cloudOut);
+	static void cvMat2Cloud(double minHeight, double maxHeight,cv::Mat& imageIn, pcl::PointCloud<pcl::PointXYZ>::Ptr cloudOut);
 	
 	/// <summary>
 	/// cv::Mat 查找轮廓
